Create controllers and services for all calculations
