+++ src/main/java/org/gridsuite/study/server/controller/StudyController.java
@@ -25,24 +25,18 @@ import org.gridsuite.study.server.dto.elasticsearch.EquipmentInfos;
 import org.gridsuite.study.server.dto.modification.*;
 import org.gridsuite.study.server.dto.networkexport.ExportNetworkStatus;
 import org.gridsuite.study.server.dto.networkexport.NodeExportInfos;
-import org.gridsuite.study.server.dto.sensianalysis.SensitivityAnalysisCsvFileInfos;
 import org.gridsuite.study.server.dto.sequence.NodeSequenceType;
 import org.gridsuite.study.server.dto.timeseries.TimeSeriesMetadataInfos;
 import org.gridsuite.study.server.dto.timeseries.TimelineEventInfos;
-import org.gridsuite.study.server.dto.voltageinit.parameters.StudyVoltageInitParameters;
 import org.gridsuite.study.server.elasticsearch.EquipmentInfosService;
 import org.gridsuite.study.server.error.StudyException;
 import org.gridsuite.study.server.exception.PartialResultException;
 import org.gridsuite.study.server.networkmodificationtree.dto.*;
 import org.gridsuite.study.server.service.*;
-import org.gridsuite.study.server.service.shortcircuit.FaultResultsMode;
-import org.gridsuite.study.server.service.shortcircuit.ShortcircuitAnalysisType;
-import org.gridsuite.study.server.utils.ResultParameters;
 import org.springframework.core.io.Resource;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.util.Pair;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -75,7 +69,6 @@ public class StudyController {
     private final RemoteServicesInspector remoteServicesInspector;
     private final RootNetworkService rootNetworkService;
     private final RootNetworkNodeInfoService rootNetworkNodeInfoService;
-    private final SensitivityAnalysisRestService sensitivityAnalysisService;
     private final RebuildNodeService rebuildNodeService;
 
     public StudyController(StudyService studyService,
@@ -87,7 +80,6 @@ public class StudyController {
                            RemoteServicesInspector remoteServicesInspector,
                            RootNetworkService rootNetworkService,
                            RootNetworkNodeInfoService rootNetworkNodeInfoService,
-                           SensitivityAnalysisRestService sensitivityAnalysisService,
                            RebuildNodeService rebuildNodeService) {
         this.studyService = studyService;
         this.networkModificationTreeService = networkModificationTreeService;
@@ -98,7 +90,6 @@ public class StudyController {
         this.remoteServicesInspector = remoteServicesInspector;
         this.rootNetworkService = rootNetworkService;
         this.rootNetworkNodeInfoService = rootNetworkNodeInfoService;
-        this.sensitivityAnalysisService = sensitivityAnalysisService;
         this.rebuildNodeService = rebuildNodeService;
     }
 
@@ -743,184 +734,6 @@ public class StudyController {
         }
     }
 
-    @PutMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/shortcircuit/run")
-    @Operation(summary = "run short circuit analysis on study")
-    @ApiResponse(responseCode = "200", description = "The short circuit analysis has started")
-    public ResponseEntity<Void> runShortCircuit(
-            @PathVariable("studyUuid") UUID studyUuid,
-            @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-            @PathVariable("nodeUuid") UUID nodeUuid,
-            @RequestParam(value = "busId", required = false) Optional<String> busId,
-            @RequestParam(name = "debug", required = false, defaultValue = "false") boolean debug,
-            @RequestHeader(HEADER_USER_ID) String userId) {
-        studyService.assertIsNodeNotReadOnly(nodeUuid);
-        studyService.assertOnQuotasAvailability(SHORT_CIRCUIT, userId);
-        studyService.runShortCircuit(studyUuid, nodeUuid, rootNetworkUuid, busId, debug, userId);
-        return ResponseEntity.ok().build();
-    }
-
-    @PutMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/shortcircuit/stop")
-    @Operation(summary = "stop security analysis on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The short circuit analysis has been stopped")})
-    public ResponseEntity<Void> stopShortCircuitAnalysis(@Parameter(description = "Study uuid") @PathVariable("studyUuid") UUID studyUuid,
-                                                         @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                         @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid,
-                                                         @RequestHeader(HEADER_USER_ID) String userId) {
-        rootNetworkNodeInfoService.stopShortCircuitAnalysis(studyUuid, nodeUuid, rootNetworkUuid, userId);
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/shortcircuit/result")
-    @Operation(summary = "Get a short circuit analysis result on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The short circuit analysis result"),
-        @ApiResponse(responseCode = "204", description = "No short circuit analysis has been done yet"),
-        @ApiResponse(responseCode = "404", description = "The short circuit analysis has not been found")})
-    public ResponseEntity<String> getShortCircuitResult(@Parameter(description = "study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                        @Parameter(description = "root network UUID") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                        @Parameter(description = "node UUID") @PathVariable("nodeUuid") UUID nodeUuid,
-                                                        @Parameter(description = "BASIC (faults without limits and feeders), " +
-                                                            "FULL (faults with both), " +
-                                                            "WITH_LIMIT_VIOLATIONS (like FULL but only those with limit violations) or " +
-                                                            "NONE (no fault)") @RequestParam(name = "mode", required = false, defaultValue = "FULL") FaultResultsMode mode,
-                                                        @Parameter(description = "type") @RequestParam(value = "type", required = false, defaultValue = "ALL_BUSES") ShortcircuitAnalysisType type,
-                                                        @Parameter(description = "JSON array of filters") @RequestParam(name = "filters", required = false) String filters,
-                                                        @Parameter(description = "JSON array of global filters") @RequestParam(name = "globalFilters", required = false) String globalFilters,
-                                                        @Parameter(description = "If we wanted the paged version of the results or not") @RequestParam(name = "paged", required = false,
-                                                                defaultValue = "false") boolean paged,
-                                                        Pageable pageable) {
-        String result = rootNetworkNodeInfoService.getShortCircuitAnalysisResult(new ResultParameters(rootNetworkUuid, nodeUuid), mode, type, filters, globalFilters, paged, pageable);
-        return result != null ? ResponseEntity.ok().body(result) :
-                ResponseEntity.noContent().build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/shortcircuit/status")
-    @Operation(summary = "Get the short circuit analysis status on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The short circuit analysis status"),
-        @ApiResponse(responseCode = "204", description = "No short circuit analysis has been done yet"),
-        @ApiResponse(responseCode = "404", description = "The short circuit analysis status has not been found")})
-    public ResponseEntity<String> getShortCircuitAnalysisStatus(@Parameter(description = "Study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                                @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                                @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid,
-                                                                @Parameter(description = "type") @RequestParam(value = "type", required = false,
-                                                                        defaultValue = "ALL_BUSES") ShortcircuitAnalysisType type) {
-        String result = rootNetworkNodeInfoService.getShortCircuitAnalysisStatus(nodeUuid, rootNetworkUuid, type);
-        return result != null ? ResponseEntity.ok().body(result) :
-                ResponseEntity.noContent().build();
-    }
-
-    @PostMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/shortcircuit/result/csv")
-    @Operation(summary = "Get a short circuit analysis csv result")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The short circuit analysis csv export"),
-        @ApiResponse(responseCode = "204", description = "No short circuit analysis has been done yet"),
-        @ApiResponse(responseCode = "404", description = "The short circuit analysis has not been found")})
-    public ResponseEntity<byte[]> getShortCircuitAnalysisCsvResult(
-            @Parameter(description = "study UUID") @PathVariable("studyUuid") UUID studyUuid,
-            @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-            @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid,
-            @Parameter(description = "type") @RequestParam(value = "type") ShortcircuitAnalysisType type,
-            @Parameter(description = "JSON array of filters") @RequestParam(name = "filters", required = false) String filters,
-            @Parameter(description = "JSON array of global filters") @RequestParam(name = "globalFilters", required = false) String globalFilters,
-            @Parameter(description = "headersCsv") @RequestBody String headersCsv,
-            Sort sort) {
-        return ResponseEntity.ok().body(rootNetworkNodeInfoService.getShortCircuitAnalysisCsvResult(nodeUuid, rootNetworkUuid, type, filters, globalFilters, sort, headersCsv));
-    }
-
-    @PostMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/pcc-min/result/csv", consumes = MediaType.APPLICATION_JSON_VALUE)
-    @Operation(summary = "Get a pcc min result as csv")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "Csv of pcc min results"),
-        @ApiResponse(responseCode = "204", description = "No pcc min has been done yet"),
-        @ApiResponse(responseCode = "404", description = "The pcc min has not been found")})
-    public ResponseEntity<byte[]> exportPccMinResultsAsCsv(
-        @Parameter(description = "study UUID") @PathVariable("studyUuid") UUID studyUuid,
-        @Parameter(description = "rootNetworkUuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-        @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid,
-        @Parameter(description = "JSON array of filters") @RequestParam(name = "filters", required = false) String filters,
-        @Parameter(description = "JSON array of global filters") @RequestParam(name = "globalFilters", required = false) String globalFilters,
-        Sort sort, @RequestBody String csvHeaders) {
-        byte[] result = rootNetworkNodeInfoService.exportPccMinResultsAsCsv(nodeUuid, rootNetworkUuid, csvHeaders, sort, filters, globalFilters);
-        HttpHeaders responseHeaders = new HttpHeaders();
-        responseHeaders.setContentType(MediaType.APPLICATION_OCTET_STREAM);
-        responseHeaders.setContentDispositionFormData("attachment", "pcc_min_results.csv");
-
-        return ResponseEntity
-            .ok()
-            .headers(responseHeaders)
-            .body(result);
-    }
-
-    @PutMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/voltage-init/run")
-    @Operation(summary = "run voltage init on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The voltage init has started"),
-        @ApiResponse(responseCode = "403", description = "The study node is not a model node")})
-    public ResponseEntity<Void> runVoltageInit(
-            @PathVariable("studyUuid") UUID studyUuid,
-            @Parameter(description = "rootNetworkUuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-            @PathVariable("nodeUuid") UUID nodeUuid,
-            @Parameter(description = "debug") @RequestParam(name = "debug", required = false, defaultValue = "false") boolean debug,
-            @RequestHeader(HEADER_USER_ID) String userId) {
-        studyService.assertIsNodeNotReadOnly(nodeUuid);
-        studyService.assertOnQuotasAvailability(VOLTAGE_INITIALIZATION, userId);
-        studyService.runVoltageInit(studyUuid, nodeUuid, rootNetworkUuid, userId, debug);
-        return ResponseEntity.ok().build();
-    }
-
-    @PutMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/voltage-init/stop")
-    @Operation(summary = "stop security analysis on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The voltage init has been stopped")})
-    public ResponseEntity<Void> stopVoltageInit(@Parameter(description = "Study uuid") @PathVariable("studyUuid") UUID studyUuid,
-                                                @Parameter(description = "rootNetworkUuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid,
-                                                @RequestHeader(HEADER_USER_ID) String userId) {
-        rootNetworkNodeInfoService.stopVoltageInit(studyUuid, nodeUuid, rootNetworkUuid, userId);
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/voltage-init/result")
-    @Operation(summary = "Get a voltage init result on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The voltage init result"),
-        @ApiResponse(responseCode = "204", description = "No voltage init has been done yet"),
-        @ApiResponse(responseCode = "404", description = "The voltage init has not been found")})
-    public ResponseEntity<String> getVoltageInitResult(@Parameter(description = "study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                        @Parameter(description = "rootNetworkUuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                        @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid,
-                                                        @Parameter(description = "JSON array of global filters") @RequestParam(name = "globalFilters", required = false) String globalFilters) {
-        String result = studyService.getVoltageInitResult(nodeUuid, rootNetworkUuid, globalFilters);
-        return result != null ? ResponseEntity.ok().body(result) :
-                ResponseEntity.noContent().build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/voltage-init/status")
-    @Operation(summary = "Get the voltage init status on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The voltage init status"),
-        @ApiResponse(responseCode = "204", description = "No voltage init has been done yet"),
-        @ApiResponse(responseCode = "404", description = "The voltage init status has not been found")})
-    public ResponseEntity<String> getVoltageInitStatus(@Parameter(description = "Study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                       @Parameter(description = "rootNetworkUuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                       @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid) {
-        String result = rootNetworkNodeInfoService.getVoltageInitStatus(nodeUuid, rootNetworkUuid);
-        return result != null ? ResponseEntity.ok().body(result) :
-                ResponseEntity.noContent().build();
-    }
-
-    @PostMapping(value = "/studies/{studyUuid}/voltage-init/parameters")
-    @Operation(summary = "Set voltage init parameters on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The voltage init parameters are set"),
-        @ApiResponse(responseCode = "204", description = "Reset with user profile cannot be done")})
-    public ResponseEntity<Void> setVoltageInitParameters(
-            @PathVariable("studyUuid") UUID studyUuid,
-            @RequestBody(required = false) StudyVoltageInitParameters voltageInitParameters,
-            @RequestHeader(HEADER_USER_ID) String userId) {
-        return studyService.setVoltageInitParameters(studyUuid, voltageInitParameters, userId) ? ResponseEntity.noContent().build() : ResponseEntity.ok().build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/voltage-init/parameters")
-    @Operation(summary = "Get voltage init parameters on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The voltage init parameters")})
-    public ResponseEntity<StudyVoltageInitParameters> getVoltageInitParameters(
-            @PathVariable("studyUuid") UUID studyUuid) {
-        return ResponseEntity.ok().body(studyService.getVoltageInitParameters(studyUuid));
-    }
-
     @GetMapping(value = "/export-network-formats")
     @Operation(summary = "get the available export format")
     @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The available export format")})
@@ -1027,24 +840,6 @@ public class StudyController {
     @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The dynamic margin calculation provider is returned")})
     public ResponseEntity<String> getDynamicMarginCalculationProvider(@PathVariable("studyUuid") UUID studyUuid) {
         return ResponseEntity.ok().body(studyService.getDynamicMarginCalculationProvider(studyUuid));
-    }
-
-    @PostMapping(value = "/studies/{studyUuid}/short-circuit-analysis/parameters", consumes = MediaType.APPLICATION_JSON_VALUE)
-    @Operation(summary = "set short-circuit analysis parameters on study, reset to default ones if empty body")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The short-circuit analysis parameters are set"),
-        @ApiResponse(responseCode = "204", description = "Reset with user profile cannot be done")})
-    public ResponseEntity<Void> setShortCircuitParameters(
-            @PathVariable("studyUuid") UUID studyUuid,
-            @RequestBody(required = false) String shortCircuitParametersInfos,
-            @RequestHeader(HEADER_USER_ID) String userId) {
-        return studyService.setShortCircuitParameters(studyUuid, shortCircuitParametersInfos, userId) ? ResponseEntity.noContent().build() : ResponseEntity.ok().build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/short-circuit-analysis/parameters", produces = MediaType.APPLICATION_JSON_VALUE)
-    @Operation(summary = "Get short-circuit analysis parameters on study")
-    @ApiResponse(responseCode = "200", description = "The short-circuit analysis parameters return by shortcircuit-server")
-    public ResponseEntity<String> getShortCircuitParameters(@PathVariable("studyUuid") UUID studyUuid) {
-        return ResponseEntity.ok().body(studyService.getShortCircuitParametersInfo(studyUuid));
     }
 
     @PostMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/network/substations/{substationId}/svg")
@@ -1577,101 +1372,6 @@ public class StudyController {
         return ResponseEntity.ok().build();
     }
 
-    @PostMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/sensitivity-analysis/run")
-    @Operation(summary = "run sensitivity analysis on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The sensitivity analysis has started"), @ApiResponse(responseCode = "403",
-            description = "The study node is not a model node")})
-    public ResponseEntity<Void> runSensitivityAnalysis(@Parameter(description = "studyUuid") @PathVariable("studyUuid") UUID studyUuid,
-                                                       @Parameter(description = "rootNetworkUuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                       @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid,
-                                                       @RequestHeader(HEADER_USER_ID) String userId) {
-        studyService.assertIsNodeNotReadOnly(nodeUuid);
-        studyService.assertOnQuotasAvailability(SENSITIVITY_ANALYSIS, userId);
-        studyService.runSensitivityAnalysis(studyUuid, nodeUuid, rootNetworkUuid, userId);
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/sensitivity-analysis/result")
-    @Operation(summary = "Get a sensitivity analysis result on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The sensitivity analysis result"),
-        @ApiResponse(responseCode = "204", description = "No sensitivity analysis has been done yet"),
-        @ApiResponse(responseCode = "404", description = "The sensitivity analysis has not been found")})
-    public ResponseEntity<String> getSensitivityAnalysisResult(
-        @Parameter(description = "study UUID") @PathVariable("studyUuid") UUID studyUuid,
-        @Parameter(description = "rootNetworkUuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-        @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid,
-        @Parameter(description = "results selector") @RequestParam("selector") String selector,
-        @Parameter(description = "JSON array of filters") @RequestParam(name = "filters", required = false) String filters,
-        @Parameter(description = "JSON array of global filters") @RequestParam(name = "globalFilters", required = false) String globalFilters
-    ) {
-        String result = rootNetworkNodeInfoService.getSensitivityAnalysisResult(nodeUuid, rootNetworkUuid, selector, filters, globalFilters);
-        return result != null ? ResponseEntity.ok().body(result) :
-            ResponseEntity.noContent().build();
-    }
-
-    @PostMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/sensitivity-analysis/result/csv", consumes = MediaType.APPLICATION_JSON_VALUE)
-    @Operation(summary = "Get a sensitivity analysis result as csv")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "Csv of sensitivity analysis results"),
-        @ApiResponse(responseCode = "204", description = "No sensitivity analysis has been done yet"),
-        @ApiResponse(responseCode = "404", description = "The sensitivity analysis has not been found")})
-    public ResponseEntity<byte[]> exportSensitivityResultsAsCsv(
-        @Parameter(description = "study UUID") @PathVariable("studyUuid") UUID studyUuid,
-        @Parameter(description = "rootNetworkUuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-        @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid,
-        @Parameter(description = "results selector") @RequestParam("selector") String selector,
-        @Parameter(description = "JSON array of filters") @RequestParam(name = "filters", required = false) String filters,
-        @Parameter(description = "JSON array of global filters") @RequestParam(name = "globalFilters", required = false) String globalFilters,
-        @RequestBody SensitivityAnalysisCsvFileInfos sensitivityAnalysisCsvFileInfos) {
-        byte[] result = rootNetworkNodeInfoService.exportSensitivityResultsAsCsv(nodeUuid, rootNetworkUuid, sensitivityAnalysisCsvFileInfos, selector, filters, globalFilters);
-        HttpHeaders responseHeaders = new HttpHeaders();
-        responseHeaders.setContentType(MediaType.APPLICATION_OCTET_STREAM);
-        responseHeaders.setContentDispositionFormData("attachment", "sensitivity_results.csv");
-
-        return ResponseEntity
-                .ok()
-                .headers(responseHeaders)
-                .body(result);
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/sensitivity-analysis/result/filter-options")
-    @Operation(summary = "Get sensitivity analysis filter options on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The sensitivity analysis filter options"),
-        @ApiResponse(responseCode = "204", description = "No sensitivity analysis has been done yet"),
-        @ApiResponse(responseCode = "404", description = "The sensitivity analysis has not been found")})
-    public ResponseEntity<String> getSensitivityAnalysisFilterOptions(
-        @Parameter(description = "study UUID") @PathVariable("studyUuid") UUID studyUuid,
-        @Parameter(description = "rootNetworkUuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-        @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid,
-        @Parameter(description = "results selector") @RequestParam("selector") String selector) {
-        String result = rootNetworkNodeInfoService.getSensitivityResultsFilterOptions(nodeUuid, rootNetworkUuid, selector);
-        return result != null ? ResponseEntity.ok().body(result) :
-            ResponseEntity.noContent().build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/sensitivity-analysis/status")
-    @Operation(summary = "Get the sensitivity analysis status on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The sensitivity analysis status"),
-        @ApiResponse(responseCode = "204", description = "No sensitivity analysis has been done yet"),
-        @ApiResponse(responseCode = "404", description = "The sensitivity analysis status has not been found")})
-    public ResponseEntity<String> getSensitivityAnalysisStatus(@Parameter(description = "Study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                               @Parameter(description = "rootNetworkUuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                               @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid) {
-        String result = rootNetworkNodeInfoService.getSensitivityAnalysisStatus(nodeUuid, rootNetworkUuid);
-        return result != null ? ResponseEntity.ok().body(result) :
-            ResponseEntity.noContent().build();
-    }
-
-    @PutMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/sensitivity-analysis/stop")
-    @Operation(summary = "stop sensitivity analysis on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The sensitivity analysis has been stopped")})
-    public ResponseEntity<Void> stopSensitivityAnalysis(@Parameter(description = "Study uuid") @PathVariable("studyUuid") UUID studyUuid,
-                                                        @Parameter(description = "rootNetworkUuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                        @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid,
-                                                        @RequestHeader(HEADER_USER_ID) String userId) {
-        rootNetworkNodeInfoService.stopSensitivityAnalysis(studyUuid, nodeUuid, rootNetworkUuid, userId);
-        return ResponseEntity.ok().build();
-    }
-
     // --- Dynamic Simulation Endpoints BEGIN --- //
 
     @PostMapping(value = "/studies/{studyUuid}/dynamic-simulation/parameters")
@@ -2015,37 +1715,6 @@ public class StudyController {
         }
     }
 
-    @GetMapping(value = "/studies/{studyUuid}/sensitivity-analysis/parameters")
-    @Operation(summary = "Get sensitivity analysis parameters on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The sensitivity analysis parameters")})
-    public ResponseEntity<String> getSensitivityAnalysisParameters(
-            @PathVariable("studyUuid") UUID studyUuid) {
-        return ResponseEntity.ok().body(studyService.getSensitivityAnalysisParameters(studyUuid));
-    }
-
-    @PostMapping(value = "/studies/{studyUuid}/sensitivity-analysis/parameters")
-    @Operation(summary = "set sensitivity analysis parameters on study, reset to default ones if empty body")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The sensitivity analysis parameters are set"),
-        @ApiResponse(responseCode = "204", description = "Reset with user profile cannot be done")})
-    public ResponseEntity<Void> setSensitivityAnalysisParameters(
-            @PathVariable("studyUuid") UUID studyUuid,
-            @RequestBody(required = false) String sensitivityAnalysisParameters,
-            @RequestHeader(HEADER_USER_ID) String userId) {
-        return studyService.setSensitivityAnalysisParameters(studyUuid, sensitivityAnalysisParameters, userId) ? ResponseEntity.noContent().build() : ResponseEntity.ok().build();
-    }
-
-    @PostMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/sensitivity-analysis/factor-count")
-    @Operation(summary = "Get the factor count of sensitivity parameters")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The factors count of sensitivity parameters")})
-    public ResponseEntity<String> getSensitivityAnalysisFactorCount(
-            @PathVariable("studyUuid") UUID studyUuid,
-            @Parameter(description = "rootNetworkUuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-            @PathVariable("nodeUuid") UUID nodeUuid,
-            @RequestBody String sensitivityAnalysisParameters) {
-        return ResponseEntity.ok().body(sensitivityAnalysisService.getSensitivityAnalysisFactorCount(rootNetworkService.getNetworkUuid(rootNetworkUuid),
-            networkModificationTreeService.getVariantId(nodeUuid, rootNetworkUuid), sensitivityAnalysisParameters));
-    }
-
     @GetMapping(value = "/servers/infos")
     @Operation(summary = "Get the information of all backend servers (if not filter with view parameter)")
     @ApiResponses(value = {
@@ -2147,127 +1816,6 @@ public class StudyController {
         return ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body(studyService.exportFilters(rootNetworkUuid, filtersUuid, nodeUuid, inUpstreamBuiltParentNode));
     }
 
-    @PostMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/state-estimation/run")
-    @Operation(summary = "run state estimation on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The state estimation has started")})
-    public ResponseEntity<Void> runStateEstimation(@Parameter(description = "studyUuid") @PathVariable("studyUuid") UUID studyUuid,
-                                                    @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                    @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid,
-                                                    @RequestParam(name = "debug", required = false, defaultValue = "false") boolean debug,
-                                                    @RequestHeader(HEADER_USER_ID) String userId) {
-        studyService.assertIsNodeNotReadOnly(nodeUuid);
-        studyService.assertOnQuotasAvailability(STATE_ESTIMATION, userId);
-        studyService.runStateEstimation(studyUuid, nodeUuid, rootNetworkUuid, userId, debug);
-        return ResponseEntity.ok().build();
-    }
-
-    @PostMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/pcc-min/run")
-    @Operation(summary = "run pcc min on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The pcc min has started")})
-    public ResponseEntity<Void> runPccMin(@Parameter(description = "studyUuid") @PathVariable("studyUuid") UUID studyUuid,
-                                                    @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                          @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid,
-                                          @RequestHeader(HEADER_USER_ID) String userId) {
-
-        studyService.assertIsNodeNotReadOnly(nodeUuid);
-        studyService.assertOnQuotasAvailability(PCC_MIN, userId);
-        studyService.runPccMin(studyUuid, nodeUuid, rootNetworkUuid, userId);
-        return ResponseEntity.ok().build();
-    }
-
-    @PutMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/pcc-min/stop")
-    @Operation(summary = "stop pcc min on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The pcc min has been stopped")})
-    public ResponseEntity<Void> stopPccMin(@Parameter(description = "Study uuid") @PathVariable("studyUuid") UUID studyUuid,
-                                                    @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                    @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid) {
-        rootNetworkNodeInfoService.stopPccMin(studyUuid, nodeUuid, rootNetworkUuid);
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/pcc-min/result")
-    @Operation(summary = "Get a pcc min result on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The pcc min result"),
-        @ApiResponse(responseCode = "204", description = "No pcc min  has been done yet"),
-        @ApiResponse(responseCode = "404", description = "The pcc min  has not been found")})
-    public ResponseEntity<String> getPccMinResult(@Parameter(description = "study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                  @Parameter(description = "rootNetwork Uuid") @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                  @Parameter(description = "node Uuid") @PathVariable("nodeUuid") UUID nodeUuid,
-                                                  @Parameter(description = "JSON array of filters") @RequestParam(name = "filters", required = false) String filters,
-                                                  @Parameter(description = "JSON array of global filters") @RequestParam(name = "globalFilters", required = false) String globalFilters,
-                                                  Pageable pageable) {
-        String result = rootNetworkNodeInfoService.getPccMinResult(nodeUuid, rootNetworkUuid, filters, globalFilters, pageable);
-        return result != null ? ResponseEntity.ok().body(result) :
-            ResponseEntity.noContent().build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/state-estimation/result")
-    @Operation(summary = "Get a state estimation result on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The state estimation result"),
-        @ApiResponse(responseCode = "204", description = "No state estimation has been done yet"),
-        @ApiResponse(responseCode = "404", description = "The state estimation has not been found")})
-    public ResponseEntity<String> getStateEstimationResult(@Parameter(description = "study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                           @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                            @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid) {
-        String result = rootNetworkNodeInfoService.getStateEstimationResult(nodeUuid, rootNetworkUuid);
-        return result != null ? ResponseEntity.ok().body(result) :
-            ResponseEntity.noContent().build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/state-estimation/status")
-    @Operation(summary = "Get the state estimation status on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The state estimation status"),
-        @ApiResponse(responseCode = "204", description = "No state estimation has been done yet"),
-        @ApiResponse(responseCode = "404", description = "The state estimation status has not been found")})
-    public ResponseEntity<String> getStateEstimationStatus(@Parameter(description = "Study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                            @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                            @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid) {
-        String status = rootNetworkNodeInfoService.getStateEstimationStatus(nodeUuid, rootNetworkUuid);
-        return status != null ? ResponseEntity.ok().body(status) : ResponseEntity.noContent().build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/pcc-min/status")
-    @Operation(summary = "Get the pcc min status on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The pcc min status"),
-        @ApiResponse(responseCode = "204", description = "No pcc min has been done yet"),
-        @ApiResponse(responseCode = "404", description = "The pcc min status has not been found")})
-    public ResponseEntity<String> getPccMinStatus(@Parameter(description = "Study UUID") @PathVariable("studyUuid") UUID studyUuid,
-                                                  @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                  @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid) {
-        String status = rootNetworkNodeInfoService.getPccMinStatus(nodeUuid, rootNetworkUuid);
-        return status != null ? ResponseEntity.ok().body(status) : ResponseEntity.noContent().build();
-    }
-
-    @PutMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/state-estimation/stop")
-    @Operation(summary = "stop state estimation on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The state estimation has been stopped")})
-    public ResponseEntity<Void> stopStateEstimation(@Parameter(description = "Study uuid") @PathVariable("studyUuid") UUID studyUuid,
-                                                     @PathVariable("rootNetworkUuid") UUID rootNetworkUuid,
-                                                     @Parameter(description = "nodeUuid") @PathVariable("nodeUuid") UUID nodeUuid) {
-        rootNetworkNodeInfoService.stopStateEstimation(studyUuid, nodeUuid, rootNetworkUuid);
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/state-estimation/parameters")
-    @Operation(summary = "Get state estimation parameters on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The state estimation parameters")})
-    public ResponseEntity<String> getStateEstimationParametersValues(
-        @PathVariable("studyUuid") UUID studyUuid) {
-        return ResponseEntity.ok().body(studyService.getStateEstimationParameters(studyUuid));
-    }
-
-    @PostMapping(value = "/studies/{studyUuid}/state-estimation/parameters")
-    @Operation(summary = "set state estimation parameters on study, reset to default ones if empty body")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The state estimation parameters are set"),
-        @ApiResponse(responseCode = "204", description = "Reset with user profile cannot be done")})
-    public ResponseEntity<Void> setStateEstimationParametersValues(
-        @PathVariable("studyUuid") UUID studyUuid,
-        @RequestBody(required = false) String stateEstimationParametersValues,
-        @RequestHeader(HEADER_USER_ID) String userId) {
-        studyService.setStateEstimationParametersValues(studyUuid, stateEstimationParametersValues, userId);
-        return ResponseEntity.ok().build();
-    }
-
     @GetMapping(value = "/studies/{studyUuid}/node-aliases")
     @Operation(summary = "Get node aliases attached to a given study")
     @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The study's attached aliases")})
@@ -2304,25 +1852,6 @@ public class StudyController {
     public ResponseEntity<Void> updateSpreadsheetParameters(@PathVariable("studyUuid") final UUID studyUuid,
                                                             @RequestBody final SpreadsheetParameters spreadsheetParameters) {
         return (this.studyService.updateSpreadsheetParameters(studyUuid, spreadsheetParameters) ? ResponseEntity.noContent() : ResponseEntity.notFound()).build();
-    }
-
-    @GetMapping(value = "/studies/{studyUuid}/pcc-min/parameters")
-    @Operation(summary = "Get pcc min parameters on study")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The pcc min parameters")})
-    public ResponseEntity<String> getPccMinParameters(
-        @PathVariable("studyUuid") UUID studyUuid) {
-        return ResponseEntity.ok().body(studyService.getPccMinParameters(studyUuid));
-    }
-
-    @PostMapping(value = "/studies/{studyUuid}/pcc-min/parameters")
-    @Operation(summary = "set pcc min parameters on study, reset to default ones if empty body")
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "The pcc min parameters are set"),
-        @ApiResponse(responseCode = "204", description = "Reset with user profile cannot be done")})
-    public ResponseEntity<Void> setPccMinParameters(
-        @PathVariable("studyUuid") UUID studyUuid,
-        @RequestBody(required = false) String pccMinParametersInfos,
-        @RequestHeader(HEADER_USER_ID) String userId) {
-        return studyService.setPccMinParameters(studyUuid, pccMinParametersInfos, userId) ? ResponseEntity.noContent().build() : ResponseEntity.ok().build();
     }
 
     @GetMapping(value = "/studies/{studyUuid}/root-networks/{rootNetworkUuid}/nodes/{nodeUuid}/computations/status")
